fix: moves CardEmptySection inside ImpactPage
