+++ src/pages/ImpactPage/CardEmptySection/index.tsx
