+++ src/pages/ImpactPage/CardEmptySection/styles.ts
